ADM-438 refactor model and mapper directory structure
