+++ frontend/src/clients/report/dto/responseDTO.ts
