+++ frontend/src/constants.ts
@@ -116,11 +116,11 @@ export enum VelocityMetricsName {
 
 export enum CycleTimeMetricsName {
   AVERAGE_CYCLE_TIME = 'Average Cycle Time',
-  DEVELOPMENT_PROPORTION = 'Total Development Time/Total Cycle Time',
-  WAITING_PROPORTION = 'Total Waiting Time/Total Cycle Time',
-  BLOCK_PROPORTION = 'Total Block Time/Total Cycle Time',
-  REVIEW_PROPORTION = 'Total Review Time/Total Cycle Time',
-  TESTING_PROPORTION = 'Total Testing Time/Total Cycle Time',
+  DEVELOPMENT_PROPORTION = 'Total Development Time / Total Cycle Time',
+  WAITING_PROPORTION = 'Total Waiting Time / Total Cycle Time',
+  BLOCK_PROPORTION = 'Total Block Time / Total Cycle Time',
+  REVIEW_PROPORTION = 'Total Review Time / Total Cycle Time',
+  TESTING_PROPORTION = 'Total Testing Time / Total Cycle Time',
   AVERAGE_DEVELOPMENT_TIME = 'Average Development Time',
   AVERAGE_WAITING_TIME = 'Average Waiting Time',
   AVERAGE_BLOCK_TIME = 'Average Block Time',
