+++ frontend/src/hooks/reportMapper/reportUIDataStructure.ts
